Implement consolidation of printing in a (asynchronous) Servitium::Service
