+++ app/jobs/papyrus/application_job.rb
@@ -1,6 +1,6 @@
 module Papyrus
   class ApplicationJob < ActiveJob::Base
     discard_on ActiveJob::DeserializationError
-     include Papyrus::Concerns::Consolidation
+    include Papyrus::Consolidation
   end
 end
